break out tickets from structs
