+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
-  "cSpell.words": ["canonicalize"]
+  "cSpell.words": ["canonicalize"],
+  "cSpell.allowCompoundWords": true
 }
